Changes in Cryst_calculator_console for allowing working when cp=dp (current precision = double precision). The 64 bits CFML_GlobalDeps_Windows_intel64.f90 has as default cp=sp (double precision does not work with variables that have to be directly on Winteracter applications)
git-svn-id: https://forge.ill.fr/svn/crysfml@885 06508225-4b59-0410-a2a9-fdab55894d57

+++ Program_Examples/Cryst_calculator_console/menu_4.f90
@@ -68,10 +68,10 @@
     !!
     Subroutine Menu_Chem_1()
        !---- Local Variables ----!
-       character(len=20)     :: line
-       integer               :: i, j, n, iv, ierr, npos
-       integer, dimension(1) :: ivet
-       real, dimension(1)    :: vet
+       character(len=20)           :: line
+       integer                     :: i, j, n, iv, ierr, npos
+       integer,       dimension(1) :: ivet
+       real(kind=cp), dimension(1) :: vet
 
        call Set_Chem_Info()
